Makefile update to include unit tests (#9)
* modified make file to have unit test command
added .idea to gitignore
fixed unit test files to not throw warnings on using gym.spaces

* added new lines of files to satisfy style requirements

* fix bug in demo script

+++ all/approximation/traces/accumulating_test.py
@@ -6,7 +6,7 @@ from all.approximation.bases.fourier import FourierBasis
 from all.approximation.traces import AccumulatingTraces
 
 
-SPACE = Box(low=0, high=1, shape=(2,))
+SPACE = Box(low=0, high=1, shape=(2,), dtype=np.float32)
 STATE = np.array([0.5, 1])
